implemented 4.1 TSP using backtracking approach

+++ src/graph/VertexEdge.h
@@ -87,7 +87,7 @@ public:
 
     Vertex *getDest() const;
 
-    int getWeight() const;
+    double getWeight() const;
 
     void setReverse(Edge *_reverse);
 };
